+++ src/server/task_manager/task_manager.ts
@@ -1,0 +1,101 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import Joi from 'joi';
+import {
+  Logger,
+  TaskDefinition,
+  TaskDictionary,
+  TaskManager,
+  TaskPool,
+  TaskStore,
+  validateTaskDefinition,
+} from './task_pool';
+import { SanitizedTaskDefinition } from './task_pool/task';
+
+function taskManagerFactory(kbnServer: any, server: any, config: any) {
+  let taskManager: TaskManager;
+  return async function getTaskManager() {
+    if (taskManager) {
+      return taskManager;
+    }
+
+    const logger = new Logger((...args) => server.log(...args));
+    const callCluster = server.plugins.elasticsearch.getCluster('admin').callWithInternalUser;
+    const numWorkers = config.get('taskManager.num_workers');
+    const store = new TaskStore({
+      index: config.get('taskManager.index'),
+      callCluster,
+      maxAttempts: config.get('taskManager.max_attempts'),
+    });
+
+    logger.debug('Initializing the task manager index');
+    await store.init();
+
+    const definitions = extractTaskDefinitions(numWorkers, kbnServer.uiExports.taskDefinitions);
+
+    const pool = new TaskPool({
+      logger,
+      callCluster,
+      numWorkers,
+      store,
+      definitions,
+      pollInterval: config.get('taskManager.poll_interval'),
+      kbnServer,
+    });
+
+    pool.start();
+
+    taskManager = new TaskManager({
+      store,
+      pool,
+    });
+
+    return taskManager;
+  };
+}
+
+export async function taskManagerMixin(kbnServer: any, server: any, config: any) {
+  server.decorate('server', 'taskManager', {
+    getTaskManager: taskManagerFactory(kbnServer, server, config),
+  });
+}
+
+// TODO, move this to a file and properly test it, validate the taskDefinition via Joi or something
+function extractTaskDefinitions(
+  numWorkers: number,
+  taskDefinitions: TaskDictionary<TaskDefinition> = {}
+): TaskDictionary<SanitizedTaskDefinition> {
+  return Object.keys(taskDefinitions).reduce(
+    (acc, type) => {
+      const rawDefinition = taskDefinitions[type];
+      rawDefinition.type = type;
+      const definition = Joi.attempt(rawDefinition, validateTaskDefinition) as TaskDefinition;
+      const workersOccupied = Math.min(numWorkers, definition.workersOccupied || 1);
+
+      acc[type] = {
+        ...definition,
+        workersOccupied,
+      };
+
+      return acc;
+    },
+    {} as TaskDictionary<SanitizedTaskDefinition>
+  );
+}
